artifactory client removed from sigleton to fs-storage
